LayoutTests/inspector/canvas/console-record-offscreen-webgl-expected.txt

Test that CanvasManager is able to record actions made to offscreen WebGL canvas contexts.


== Running test suite: Canvas.recordingOffscreenWebGL
-- Running test case: Canvas.recordingOffscreenWebGL.Console
PASS: The recording should have the name "TEST".
PASS: The recording should have one frame.
PASS: The first frame should have one action.
